requirements.txt
Django==1.7.6
django-autocomplete-light==2.0.9
six==1.9.0
